+++ src/allotropy/parsers/thermo_fisher_qubit_flex/constants.py
@@ -1,0 +1,16 @@
+""" Constants file for Thermo Fisher Qubit Flex Parser"""
+
+# Instrument constants
+SOFTWARE_NAME = "Qubit Flex software"
+MODEL_NUMBER = "Qubit Flex"
+PRODUCT_MANUFACTURER = "Thermo Fisher Scientific"
+BRAND_NAME = "Qubit"
+DEVICE_TYPE = "fluorescence detector"
+DISPLAY_NAME = "Thermo Fisher Qubit Flex"
+
+# Error constants
+# Value to use for unsupported file format for the given file for the Adapter
+UNSUPPORTED_FILE_FORMAT_ERROR = (
+    "Unsupported file format. Expected xlsx or csv file. Actual: "
+)
+VALUE_ERROR = "Unable to find value for column"